Flux tests fix
