+++ tests/pipelines/controlnet_flux/test_controlnet_flux_img2img.py
@@ -3,33 +3,31 @@ import unittest
 
 import numpy as np
 import torch
-from transformers import CLIPTextConfig, CLIPTextModel, CLIPTokenizer, T5EncoderModel, T5TokenizerFast
+from transformers import AutoTokenizer, CLIPTextConfig, CLIPTextModel, CLIPTokenizer, T5EncoderModel
 
 from diffusers import (
     AutoencoderKL,
     FlowMatchEulerDiscreteScheduler,
     FluxControlNetImg2ImgPipeline,
+    FluxControlNetModel,
     FluxTransformer2DModel,
 )
-from diffusers.models import FluxControlNetModel
-from diffusers.utils import load_image
 from diffusers.utils.testing_utils import (
-    enable_full_determinism,
+    numpy_cosine_similarity_distance,
     require_torch_gpu,
     slow,
     torch_device,
 )
-from diffusers.utils.torch_utils import randn_tensor
 
-from ..test_pipelines_common import PipelineTesterMixin
-
-
-enable_full_determinism()
+from ..test_pipelines_common import (
+    PipelineTesterMixin,
+    check_qkv_fusion_matches_attn_procs_length,
+    check_qkv_fusion_processors_exist,
+)
 
 
 class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMixin):
     pipeline_class = FluxControlNetImg2ImgPipeline
-
     params = frozenset(
         [
             "prompt",
@@ -39,18 +37,20 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
             "width",
             "strength",
             "guidance_scale",
-            "num_inference_steps",
+            "controlnet_conditioning_scale",
             "prompt_embeds",
             "pooled_prompt_embeds",
         ]
     )
     batch_params = frozenset(["prompt", "image", "control_image"])
 
+    test_xformers_attention = False
+
     def get_dummy_components(self):
         torch.manual_seed(0)
         transformer = FluxTransformer2DModel(
             patch_size=1,
-            in_channels=16,
+            in_channels=4,
             num_layers=1,
             num_single_layers=1,
             attention_head_dim=16,
@@ -59,20 +59,6 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
             pooled_projection_dim=32,
             axes_dims_rope=[4, 4, 8],
         )
-
-        torch.manual_seed(0)
-        controlnet = FluxControlNetModel(
-            patch_size=1,
-            in_channels=16,
-            num_layers=1,
-            num_single_layers=1,
-            attention_head_dim=16,
-            num_attention_heads=2,
-            joint_attention_dim=32,
-            pooled_projection_dim=32,
-            axes_dims_rope=[4, 4, 8],
-        )
-
         clip_text_encoder_config = CLIPTextConfig(
             bos_token_id=0,
             eos_token_id=2,
@@ -86,6 +72,7 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
             hidden_act="gelu",
             projection_dim=32,
         )
+
         torch.manual_seed(0)
         text_encoder = CLIPTextModel(clip_text_encoder_config)
 
@@ -93,7 +80,7 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
         text_encoder_2 = T5EncoderModel.from_pretrained("hf-internal-testing/tiny-random-t5")
 
         tokenizer = CLIPTokenizer.from_pretrained("hf-internal-testing/tiny-random-clip")
-        tokenizer_2 = T5TokenizerFast.from_pretrained("hf-internal-testing/tiny-random-t5")
+        tokenizer_2 = AutoTokenizer.from_pretrained("hf-internal-testing/tiny-random-t5")
 
         torch.manual_seed(0)
         vae = AutoencoderKL(
@@ -102,12 +89,24 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
             out_channels=3,
             block_out_channels=(4,),
             layers_per_block=1,
-            latent_channels=4,
+            latent_channels=1,
             norm_num_groups=1,
             use_quant_conv=False,
             use_post_quant_conv=False,
             shift_factor=0.0609,
             scaling_factor=1.5035,
+        )
+
+        torch.manual_seed(0)
+        controlnet = FluxControlNetModel(
+            in_channels=4,
+            num_layers=1,
+            num_single_layers=1,
+            attention_head_dim=16,
+            num_attention_heads=2,
+            joint_attention_dim=32,
+            pooled_projection_dim=32,
+            axes_dims_rope=[4, 4, 8],
         )
 
         scheduler = FlowMatchEulerDiscreteScheduler()
@@ -129,19 +128,8 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
         else:
             generator = torch.Generator(device="cpu").manual_seed(seed)
 
-        image = randn_tensor(
-            (1, 3, 32, 32),
-            generator=generator,
-            device=torch.device(device),
-            dtype=torch.float32,
-        )
-
-        control_image = randn_tensor(
-            (1, 3, 32, 32),
-            generator=generator,
-            device=torch.device(device),
-            dtype=torch.float32,
-        )
+        image = torch.randn(1, 3, 32, 32).to(device)
+        control_image = torch.randn(1, 3, 32, 32).to(device)
 
         inputs = {
             "prompt": "A painting of a squirrel eating a burger",
@@ -149,85 +137,99 @@ class FluxControlNetImg2ImgPipelineFastTests(unittest.TestCase, PipelineTesterMi
             "control_image": control_image,
             "generator": generator,
             "num_inference_steps": 2,
-            "guidance_scale": 3.5,
-            "height": 8,
-            "width": 8,
-            "max_sequence_length": 48,
+            "guidance_scale": 5.0,
+            "controlnet_conditioning_scale": 1.0,
             "strength": 0.8,
+            "height": 32,
+            "width": 32,
+            "max_sequence_length": 48,
             "output_type": "np",
-            "controlnet_conditioning_scale": 0.5,
         }
-
         return inputs
 
-    def test_controlnet_img2img_flux(self):
+    def test_flux_controlnet_different_prompts(self):
+        pipe = self.pipeline_class(**self.get_dummy_components()).to(torch_device)
+
+        inputs = self.get_dummy_inputs(torch_device)
+        output_same_prompt = pipe(**inputs).images[0]
+
+        inputs = self.get_dummy_inputs(torch_device)
+        inputs["prompt_2"] = "a different prompt"
+        output_different_prompts = pipe(**inputs).images[0]
+
+        max_diff = np.abs(output_same_prompt - output_different_prompts).max()
+
+        assert max_diff > 1e-6
+
+    def test_flux_controlnet_prompt_embeds(self):
+        pipe = self.pipeline_class(**self.get_dummy_components()).to(torch_device)
+        inputs = self.get_dummy_inputs(torch_device)
+
+        output_with_prompt = pipe(**inputs).images[0]
+
+        inputs = self.get_dummy_inputs(torch_device)
+        prompt = inputs.pop("prompt")
+
+        (prompt_embeds, pooled_prompt_embeds, text_ids) = pipe.encode_prompt(
+            prompt,
+            prompt_2=None,
+            device=torch_device,
+            max_sequence_length=inputs["max_sequence_length"],
+        )
+        output_with_embeds = pipe(
+            prompt_embeds=prompt_embeds,
+            pooled_prompt_embeds=pooled_prompt_embeds,
+            **inputs,
+        ).images[0]
+
+        max_diff = np.abs(output_with_prompt - output_with_embeds).max()
+        assert max_diff < 1e-4
+
+    def test_fused_qkv_projections(self):
+        device = "cpu"  # ensure determinism for the device-dependent torch.Generator
         components = self.get_dummy_components()
-        flux_pipe = FluxControlNetImg2ImgPipeline(**components)
-        flux_pipe = flux_pipe.to(torch_device, dtype=torch.float32)
-        flux_pipe.set_progress_bar_config(disable=None)
+        pipe = self.pipeline_class(**components)
+        pipe = pipe.to(device)
+        pipe.set_progress_bar_config(disable=None)
 
-        inputs = self.get_dummy_inputs(torch_device)
-        output = flux_pipe(**inputs)
-        image = output.images
+        inputs = self.get_dummy_inputs(device)
+        image = pipe(**inputs).images
+        original_image_slice = image[0, -3:, -3:, -1]
 
-        image_slice = image[0, -3:, -3:, -1]
+        pipe.transformer.fuse_qkv_projections()
+        assert check_qkv_fusion_processors_exist(
+            pipe.transformer
+        ), "Something wrong with the fused attention processors. Expected all the attention processors to be fused."
+        assert check_qkv_fusion_matches_attn_procs_length(
+            pipe.transformer, pipe.transformer.original_attn_processors
+        ), "Something wrong with the attention processors concerning the fused QKV projections."
 
-        assert image.shape == (1, 32, 32, 3)
+        inputs = self.get_dummy_inputs(device)
+        image = pipe(**inputs).images
+        image_slice_fused = image[0, -3:, -3:, -1]
 
-        expected_slice = np.array([0.5182, 0.4976, 0.4718, 0.5249, 0.5039, 0.4751, 0.5168, 0.4980, 0.4738])
+        pipe.transformer.unfuse_qkv_projections()
+        inputs = self.get_dummy_inputs(device)
+        image = pipe(**inputs).images
+        image_slice_disabled = image[0, -3:, -3:, -1]
 
-        assert (
-            np.abs(image_slice.flatten() - expected_slice).max() < 1e-2
-        ), f"Expected: {expected_slice}, got: {image_slice.flatten()}"
-
-    def test_attention_slicing_forward_pass(self):
-        components = self.get_dummy_components()
-        flux_pipe = FluxControlNetImg2ImgPipeline(**components)
-        flux_pipe = flux_pipe.to(torch_device, dtype=torch.float32)
-        flux_pipe.set_progress_bar_config(disable=None)
-
-        flux_pipe.enable_attention_slicing()
-        inputs = self.get_dummy_inputs(torch_device)
-        output_sliced = flux_pipe(**inputs)
-        image_sliced = output_sliced.images
-
-        flux_pipe.disable_attention_slicing()
-        inputs = self.get_dummy_inputs(torch_device)
-        output = flux_pipe(**inputs)
-        image = output.images
-
-        assert np.abs(image_sliced.flatten() - image.flatten()).max() < 1e-3
-
-    def test_inference_batch_single_identical(self):
-        components = self.get_dummy_components()
-        flux_pipe = FluxControlNetImg2ImgPipeline(**components)
-        flux_pipe = flux_pipe.to(torch_device, dtype=torch.float32)
-        flux_pipe.set_progress_bar_config(disable=None)
-
-        inputs = self.get_dummy_inputs(torch_device)
-
-        # make batch size 1
-        inputs["prompt"] = [inputs["prompt"]]
-        inputs["image"] = inputs["image"][:1]
-        inputs["control_image"] = inputs["control_image"][:1]
-
-        output = flux_pipe(**inputs)
-        image = output.images
-
-        inputs["prompt"] = inputs["prompt"] * 2
-        inputs["image"] = torch.cat([inputs["image"], inputs["image"]])
-        inputs["control_image"] = torch.cat([inputs["control_image"], inputs["control_image"]])
-
-        output_batch = flux_pipe(**inputs)
-        image_batch = output_batch.images
-
-        assert np.abs(image_batch[0].flatten() - image[0].flatten()).max() < 1e-3
-        assert np.abs(image_batch[1].flatten() - image[0].flatten()).max() < 1e-3
+        assert np.allclose(
+            original_image_slice, image_slice_fused, atol=1e-3, rtol=1e-3
+        ), "Fusion of QKV projections shouldn't affect the outputs."
+        assert np.allclose(
+            image_slice_fused, image_slice_disabled, atol=1e-3, rtol=1e-3
+        ), "Outputs, with QKV projection fusion enabled, shouldn't change when fused QKV projections are disabled."
+        assert np.allclose(
+            original_image_slice, image_slice_disabled, atol=1e-2, rtol=1e-2
+        ), "Original outputs should match when fused QKV projections are disabled."
 
 
 @slow
 @require_torch_gpu
 class FluxControlNetImg2ImgPipelineSlowTests(unittest.TestCase):
+    pipeline_class = FluxControlNetImg2ImgPipeline
+    repo_id = "black-forest-labs/FLUX.1-schnell"
+
     def setUp(self):
         super().setUp()
         gc.collect()
@@ -238,80 +240,52 @@ class FluxControlNetImg2ImgPipelineSlowTests(unittest.TestCase):
         gc.collect()
         torch.cuda.empty_cache()
 
-    def test_canny(self):
-        controlnet = FluxControlNetModel.from_pretrained(
-            "InstantX/FLUX.1-dev-Controlnet-Canny-alpha", torch_dtype=torch.bfloat16
-        )
-        pipe = FluxControlNetImg2ImgPipeline.from_pretrained(
-            "black-forest-labs/FLUX.1-dev", controlnet=controlnet, torch_dtype=torch.bfloat16
-        )
+    def get_inputs(self, device, seed=0):
+        if str(device).startswith("mps"):
+            generator = torch.manual_seed(seed)
+        else:
+            generator = torch.Generator(device="cpu").manual_seed(seed)
+
+        image = torch.randn(1, 3, 64, 64).to(device)
+        control_image = torch.randn(1, 3, 64, 64).to(device)
+
+        return {
+            "prompt": "A photo of a cat",
+            "image": image,
+            "control_image": control_image,
+            "num_inference_steps": 2,
+            "guidance_scale": 5.0,
+            "controlnet_conditioning_scale": 1.0,
+            "strength": 0.8,
+            "output_type": "np",
+            "generator": generator,
+        }
+
+    @unittest.skip("We cannot run inference on this model with the current CI hardware")
+    def test_flux_controlnet_img2img_inference(self):
+        pipe = self.pipeline_class.from_pretrained(self.repo_id, torch_dtype=torch.bfloat16)
         pipe.enable_model_cpu_offload()
-        pipe.set_progress_bar_config(disable=None)
 
-        generator = torch.Generator(device="cpu").manual_seed(0)
-        prompt = "A girl in city, 25 years old, cool, futuristic"
-        control_image = load_image(
-            "https://huggingface.co/InstantX/FLUX.1-dev-Controlnet-Canny-alpha/resolve/main/canny.jpg"
-        )
-        init_image = load_image(
-            "https://huggingface.co/InstantX/FLUX.1-dev-Controlnet-Canny-alpha/resolve/main/init_image.png"
-        )
+        inputs = self.get_inputs(torch_device)
 
-        output = pipe(
-            prompt,
-            image=init_image,
-            control_image=control_image,
-            controlnet_conditioning_scale=0.6,
-            strength=0.7,
-            num_inference_steps=3,
-            guidance_scale=3.5,
-            output_type="np",
-            generator=generator,
+        image = pipe(**inputs).images[0]
+        image_slice = image[0, :10, :10]
+        expected_slice = np.array(
+            [
+                [0.36132812, 0.30004883, 0.25830078],
+                [0.36669922, 0.31103516, 0.23754883],
+                [0.34814453, 0.29248047, 0.23583984],
+                [0.35791016, 0.30981445, 0.23999023],
+                [0.36328125, 0.31274414, 0.2607422],
+                [0.37304688, 0.32177734, 0.26171875],
+                [0.3671875, 0.31933594, 0.25756836],
+                [0.36035156, 0.31103516, 0.2578125],
+                [0.3857422, 0.33789062, 0.27563477],
+                [0.3701172, 0.31982422, 0.265625],
+            ],
+            dtype=np.float32,
         )
 
-        image = output.images[0]
+        max_diff = numpy_cosine_similarity_distance(expected_slice.flatten(), image_slice.flatten())
 
-        assert image.shape == (1024, 1024, 3)
-
-        image_slice = image[-3:, -3:, -1].flatten()
-        expected_slice = np.array([0.3242, 0.3320, 0.3359, 0.3281, 0.3398, 0.3359, 0.3086, 0.3203, 0.3203])
-        assert np.abs(image_slice - expected_slice).max() < 1e-2
-
-    def test_depth(self):
-        controlnet = FluxControlNetModel.from_pretrained(
-            "InstantX/FLUX.1-dev-Controlnet-Depth-alpha", torch_dtype=torch.bfloat16
-        )
-        pipe = FluxControlNetImg2ImgPipeline.from_pretrained(
-            "black-forest-labs/FLUX.1-dev", controlnet=controlnet, torch_dtype=torch.bfloat16
-        )
-        pipe.enable_model_cpu_offload()
-        pipe.set_progress_bar_config(disable=None)
-
-        generator = torch.Generator(device="cpu").manual_seed(0)
-        prompt = "An astronaut riding a horse on mars"
-        control_image = load_image(
-            "https://huggingface.co/InstantX/FLUX.1-dev-Controlnet-Depth-alpha/resolve/main/depth.png"
-        )
-        init_image = load_image(
-            "https://huggingface.co/InstantX/FLUX.1-dev-Controlnet-Depth-alpha/resolve/main/astronaut_riding_horse.png"
-        )
-
-        output = pipe(
-            prompt,
-            image=init_image,
-            control_image=control_image,
-            controlnet_conditioning_scale=0.6,
-            strength=0.7,
-            num_inference_steps=3,
-            guidance_scale=3.5,
-            output_type="np",
-            generator=generator,
-        )
-
-        image = output.images[0]
-
-        assert image.shape == (1024, 1024, 3)
-
-        image_slice = image[-3:, -3:, -1].flatten()
-        expected_slice = np.array([0.3164, 0.3242, 0.3281, 0.3203, 0.3320, 0.3281, 0.3008, 0.3125, 0.3125])
-        assert np.abs(image_slice - expected_slice).max() < 1e-2
+        assert max_diff < 1e-4
